vk_chat_bot-master/requirements.txt
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.3
enum34==1.1.6
idna==2.8
Pillow==6.1.0
pony==0.7.10
psycopg2==2.8.3
requests==2.22.0
six==1.12.0
urllib3==1.25.3
vk-api==11.4.0
